+++ apps/new/widget/Index.jsx
@@ -2,7 +2,7 @@ const { CSS } = VM.require("${config_account}/widget/CSS") || {
   CSS: () => <></>,
 };
 
-const { Footer } = VM.require("${config_account}/widget/Components.Footer") || {
+const { Footer } = VM.require("${config_account}/widget/components.Footer") || {
   Footer: () => <></>,
 };
 
@@ -22,7 +22,7 @@ const config = {
     param: "page",
     routes: {
       home: {
-        path: "${config_account}/widget/Page.Home.Index",
+        path: "${config_account}/widget/page.home.Index",
         blockHeight: "final",
         init: {
           name: "Home",
@@ -30,7 +30,7 @@ const config = {
         default: true,
       },
       projects: {
-        path: "${config_account}/widget/Page.Projects.Index",
+        path: "${config_account}/widget/page.projects.Index",
         blockHeight: "final",
         init: {
           name: "Projects",
