+++ apps/new/widget/components/Footer.jsx
