+++ apps/new/widget/components/Sidebar.jsx
