+++ apps/new/widget/page/home/hero/ScrollingText.jsx
