+++ apps/new/widget/page/home/hero/Vertical.jsx
